Fix: 게시판 삭제 js 경로 오류 수정, Feat: 게시판 첨부파일 업로드, 다운로드 기능 추가

+++ src/main/java/org/study/board/service/BoardServiceImpl.java
@@ -3,12 +3,20 @@ package org.study.board.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+import org.study.board.domain.BoardAttachmentVO;
+import org.study.board.domain.BoardVO;
 import org.study.board.dto.BoardDTO;
 import org.study.board.mapper.BoardMapper;
+import org.study.common.util.UploadFiles;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
+
+import static org.study.common.util.UploadFiles.upload;
 
 @Log4j
 @Service
@@ -16,6 +24,7 @@ import java.util.Optional;
 public class BoardServiceImpl implements BoardService{
 
     final private BoardMapper mapper;
+    private final static String BASE_DIR = "c:/upload/springboard";
 
     @Override
     public List<BoardDTO> getList() {
@@ -34,10 +43,31 @@ public class BoardServiceImpl implements BoardService{
                 .orElseThrow(NoSuchElementException::new);
     }
 
+    @Transactional
     @Override
     public void create(BoardDTO board) {
         log.info("create........." + board);
-        mapper.create(board.toVo());
+        BoardVO boardVO = board.toVo();
+        mapper.create(boardVO);
+
+        List<MultipartFile> files = board.getFiles();
+        if (files != null && !files.isEmpty()) {
+            upload(boardVO.getNo(), files);
+        }
+    }
+
+    // 첨부파일 upload
+    private void upload(Long bno, List<MultipartFile> files) {
+        for(MultipartFile part: files) {
+            if(part.isEmpty()) continue;
+            try {
+                String uploadPath = UploadFiles.upload(BASE_DIR, part);
+                BoardAttachmentVO attach = BoardAttachmentVO.of(part, bno, uploadPath);
+                mapper.createAttachment(attach);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
     }
 
     @Override
@@ -48,7 +78,22 @@ public class BoardServiceImpl implements BoardService{
 
     @Override
     public boolean delete(Long no) {
+        BoardVO board = mapper.get(no);
+        if (board == null) {
+            return false;
+        }
         log.info("delete.........." + no);
         return mapper.delete(no) == 1;
+    }
+
+    // 첨부파일 1개 얻기
+    @Override
+    public BoardAttachmentVO getAttachment(Long no) {
+        return mapper.getAttachment(no);
+    }
+
+    @Override
+    public boolean deleteAttachment(Long no) {
+        return mapper.deleteAttachment(no) == 1;
     }
 }
